:truck: rename directory
