+++ src/getOrCreateObject/resizeAndSave.js
